FINERACT-2081: Fix delinquent data swagger type

+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/api/LoansApiResourceSwagger.java
@@ -640,9 +640,9 @@ final class LoansApiResourceSwagger {
             @Schema(example = "0.000000")
             public Double totalRepaymentTransactionReversed;
             @Schema(example = "0.000000")
-            public Double totalUnpaidAccruedDueInterest;
+            public BigDecimal totalUnpaidAccruedDueInterest;
             @Schema(example = "0.000000")
-            public Double totalUnpaidAccruedNotDueInterest;
+            public BigDecimal totalUnpaidAccruedNotDueInterest;
             public Set<GetLoansLoanIdOverdueCharges> overdueCharges;
             @Schema(example = "1")
             public Long chargeOffReasonId;
@@ -1011,13 +1011,13 @@ final class LoansApiResourceSwagger {
             @Schema(example = "100.000000")
             public Double delinquentAmount;
             @Schema(example = "80.000000")
-            public Double delinquentPrincipal;
+            public BigDecimal delinquentPrincipal;
             @Schema(example = "10.000000")
-            public Double delinquentInterest;
+            public BigDecimal delinquentInterest;
             @Schema(example = "6.000000")
-            public Double delinquentFee;
+            public BigDecimal delinquentFee;
             @Schema(example = "4.000000")
-            public Double delinquentPenalty;
+            public BigDecimal delinquentPenalty;
             @Schema(example = "[2022, 07, 01]")
             public LocalDate lastPaymentDate;
             @Schema(example = "100.000000")
